move api test fixtures to conftest so they can be used in other test modules
